% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addChildToRoot}
\alias{addChildToRoot}
\title{Add a child root (lateral) to an existing root}
\usage{
addChildToRoot(current, child)
}
\arguments{
\item{current}{= the current root}

\item{child}{= the child root to attach}
}
\value{
the current root, with the additional child attached
}
\description{
Add a child root (lateral) to an existing root
}
\examples{
data(lupin)
current <- lupin$roots[[1]]
child <- current$children[[1]]
current <- addChildToRoot(current, child)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

