% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{append.rsml}
\alias{append.rsml}
\title{Append 2 rsml files}
\usage{
append.rsml(rsml1, rsml2)
}
\arguments{
\item{rsml1}{The first RSML object. Mandatory.}

\item{rsml2}{The second RSML object. Mandatory.}
}
\description{
Append 2 rsml files
}
\examples{
append.rsml()
}
\keyword{rsml}

