% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coords}
\alias{coords}
\title{Get the coordinates of the root nodes}
\usage{
coords(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
a dataframe containing the node coordinates (x, y, z)
}
\description{
Get the coordinates of the root nodes
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{<-}
\keyword{coords(r)}
\keyword{data(lupin)}
\keyword{lupin$roots[[1]]}
\keyword{r}
\keyword{rsml}

