% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{coords.root}
\alias{coords.root}
\title{Get the coordinates of the root nodes}
\usage{
\method{coords}{root}(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
a dataframe containing the node coordinates (x, y, z)
}
\description{
Get the coordinates of the root nodes
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

