% Generated by roxygen2 (4.0.2): do not edit by hand
\name{latLength}
\alias{latLength}
\title{Compute the length of the lateral root based on the coordinates of its nodes}
\usage{
latLength(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the total length of the lateral roots
}
\description{
Compute the length of the lateral root based on the coordinates of its nodes
}
\examples{
data(lupin)
latLength(lupin)
}
\keyword{rsml}

