% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{length.plant}
\alias{length.plant}
\title{Compute the length of the root based on the coordinates of its nodes}
\usage{
\method{length}{plant}(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the total length of the plant roots
}
\description{
Compute the length of the root based on the coordinates of its nodes
}
\keyword{rsml}

