% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{length.root}
\alias{length.root}
\title{Compute the length of the root based on the coordinatres of its nodes}
\usage{
\method{length}{root}(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
the length of the root
}
\description{
Compute the length of the root based on the coordinatres of its nodes
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

