% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{meanInsertionAngle.plant}
\alias{meanInsertionAngle.plant}
\title{Compute the mean insertion angle of all the laterals in the plant}
\usage{
\method{meanInsertionAngle}{plant}(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the mean insertion angle of the root system
}
\description{
Compute the mean insertion angle of all the laterals in the plant
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

