% Generated by roxygen2 (4.0.2): do not edit by hand
\name{meanInsertionAngle}
\alias{meanInsertionAngle}
\title{Compute the mean insertion angle of the children (lateral) roots}
\usage{
meanInsertionAngle(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
the mean lateral angle
}
\description{
Compute the mean insertion angle of the children (lateral) roots
}
\examples{
data(lupin)
r <- lupin$roots[[1]]
meanInsertionAngle(r)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

