% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{meanInsertionAngle.root}
\alias{meanInsertionAngle.root}
\title{Compute the mean insertion angle of the children (lateral) roots}
\usage{
\method{meanInsertionAngle}{root}(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
the mean lateral angle
}
\description{
Compute the mean insertion angle of the children (lateral) roots
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

