% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{meanInterbranch.plant}
\alias{meanInterbranch.plant}
\title{Compute the mean interbranch distance of all the primary roots in the image}
\usage{
\method{meanInterbranch}{plant}(obj, allroot = F)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the mean interbranch distance of the root system
}
\description{
Compute the mean interbranch distance of all the primary roots in the image
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

