% Generated by roxygen2 (4.0.2): do not edit by hand
\name{meanInterbranch}
\alias{meanInterbranch}
\title{Compute the mean interbranch distance of the children (lateral) roots}
\usage{
meanInterbranch(obj, allroot = F)
}
\arguments{
\item{obj}{of class root}

\item{allroot}{if true, compute the interbanch distance on the whole root}
}
\value{
the mean interbranch distance
}
\description{
Compute the mean interbranch distance of the children (lateral) roots
}
\examples{
data(lupin)
r <- lupin$roots[[1]]
meanInterbranch(r)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

