% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{meanInterbranch.root}
\alias{meanInterbranch.root}
\title{Compute the mean interbranch distance of the children (lateral) roots}
\usage{
\method{meanInterbranch}{root}(obj, allroot = F)
}
\arguments{
\item{obj}{of class root}
}
\value{
the mean interbranch distance
}
\description{
Compute the mean interbranch distance of the children (lateral) roots
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

