% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nChild}
\alias{nChild}
\title{Get the number of children in a root}
\usage{
nChild(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
the number of child root in the current root
}
\description{
Get the number of children in a root
}
\examples{
data(lupin)
r <- lupin$roots[[1]]
nChild(r)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

