% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{nChild.root}
\alias{nChild.root}
\title{Get the number of children in a root}
\usage{
\method{nChild}{root}(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
the number of child root in the current root
}
\description{
Get the number of children in a root
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

