% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{nLatRoot.plant}
\alias{nLatRoot.plant}
\title{Compute the number of lateral roots in the plant}
\usage{
\method{nLatRoot}{plant}(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the number of lateral root in the plant
}
\description{
Compute the number of lateral roots in the plant
}
\keyword{rsml}

