% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{nNode.root}
\alias{nNode.root}
\title{Get the number of nodes in a root}
\usage{
\method{nNode}{root}(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
the number of nodes in the root
}
\description{
Get the number of nodes in a root
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

