% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{node}
\alias{node}
\title{Node constructor}
\usage{
node(x, y, z = 0, diameter = 0, orientation = 0, bLength = 0)
}
\arguments{
\item{x}{= x coordinate of the node. Mandatory}

\item{y}{= y coordinate of the node. Mandatory}

\item{z}{= z coordinate of the root. Optional}

\item{diameter}{= diameter of the node. Optional}

\item{orientation}{= orientation of the node. Optional}

\item{bLength}{= lenght from the node position in the root from the base of the root. Optional}
}
\value{
the node
}
\description{
Node constructor
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

