% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{nPrimRoot.plant}
\alias{nPrimRoot.plant}
\title{Compute the number of primary roots in the plant}
\usage{
\method{nPrimRoot}{plant}(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the number of primary root in the plant
}
\description{
Compute the number of primary roots in the plant
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

