% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nPrimRoot}
\alias{nPrimRoot}
\title{Compute the number of primary roots in the plant}
\usage{
nPrimRoot(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the number of primary root in the plant
}
\description{
Compute the number of primary roots in the plant
}
\examples{
data(lupin)
nPrimRoot(lupin)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

