% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{nRoot.plant}
\alias{nRoot.plant}
\title{Compute the total number of roots in the plant}
\usage{
\method{nRoot}{plant}(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the number of root in the plant
}
\description{
Compute the total number of roots in the plant
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

