% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{oneRSMLToList}
\alias{oneRSMLToList}
\title{Import a single RSML file into a List}
\usage{
oneRSMLToList(rsml.path, store.raw = F)
}
\arguments{
\item{rsml.path}{The path to the .rsml file. Mandatory.}

\item{store.raw}{Do you wan to store the raw RSML data into the rsml object? Take much more space. Defaults to FALSE}
}
\description{
Import a single RSML file into a List
}
\examples{
oneRSMLToList()
}
\keyword{rsml}

