% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{plot.plant}
\alias{plot.plant}
\title{Plot the root system}
\usage{
\method{plot}{plant}(obj, threed = F)
}
\arguments{
\item{obj}{of class plant}

\item{threed}{make a 3D plot for the plant}
}
\value{
null
}
\description{
Plot the root system
}
\keyword{rsml}

