% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_functions.R
\name{plot_field}
\alias{plot_field}
\title{Plot all the root system in a list of "plant" objects}
\usage{
plot_field(..., threed = F, lims = NULL, aspect = NULL)
}
\arguments{
\item{...}{plant objects, or a single list of plant objects}

\item{threed}{make a 3D plot for the plant}

\item{x}{object of class plant}
}
\value{
null
}
\description{
Plot all the root system in a list of "plant" objects
}
\keyword{rsml}
