% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.node}
\alias{print.node}
\title{Print the node}
\usage{
\method{print}{node}(x, ...)
}
\arguments{
\item{x}{object of class node}

\item{...}{print options}
}
\description{
Print the node
}
\examples{
n <- node(1, 1)
print(n)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}

