% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.plant}
\alias{print.plant}
\title{Print the plant}
\usage{
\method{print}{plant}(x, ...)
}
\arguments{
\item{x}{object of class node}

\item{...}{print options}
}
\description{
Print the plant
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}

