% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{root}
\alias{root}
\title{Root constructor}
\usage{
root(nodes = NULL, parent = "", children = NULL, id = "",
  insertion = NULL, insertion_angle = NULL)
}
\arguments{
\item{nodes}{= the nodes composing the root. Can be null}

\item{parent}{= the identifier of the root parent. Can be null}

\item{children}{= vector of children roots (root objects)}

\item{id}{= root unique identifier}

\item{insertion}{= insertion position of the root on its parent. Can be null is no parent}

\item{insertion_angle}{= insertion angle of the root on its parent. Can be null is no parent}
}
\value{
the root
}
\description{
Root constructor
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}

