\name{RSML-package}
\alias{RSML-package}
\alias{RSML}
\docType{package}
\title{
Read and analyse RSML files
}
\description{
Usefull functions to read, load and analyse RSML (Root System Markup Language) files
}
\details{
\tabular{ll}{
Package: \tab RSML\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-22\cr
License: \tab Open Source\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Guillaume Lobet

Maintainer: Guillaume Lobet <guillaume.lobet@ulg.ac.be>
}
\references{
~~ Coming Soon ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ rsml }
\seealso{
\link[http://rootsystemml.github.io/]{RSML webpage}
}
\examples{
~~ simple examples of the most important functions ~~
}
