% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{rsmlToList}
\alias{rsmlToList}
\title{Import a bunch of RSML file into a List}
\usage{
rsmlToList(all.path, store.raw = F)
}
\arguments{
\item{store.raw}{Do you wan to store the raw RSML data into the rsml object? Take much more space. Defaults to FALSE}

\item{rsml.path}{The path to the directory containing all the rsml files. Mandatory.}
}
\description{
Import a bunch of RSML file into a List
}
\examples{
rsmlToList()
}
\keyword{rsml}

