% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rsmlToPlant}
\alias{rsmlToPlant}
\title{Import a single RSML file into a Plant object}
\usage{
rsmlToPlant(rsml.path, threed = FALSE)
}
\arguments{
\item{rsml.path}{The path to the .rsml file. Mandatory.}

\item{threed}{Does the RSML file contains a 3D root system?}
}
\description{
Import a single RSML file into a Plant object
}
\keyword{3d}
\keyword{in}
\keyword{is}
\keyword{plant}
\keyword{rsml}
\keyword{the}
\keyword{threed}

