% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{summary.plant}
\alias{summary.plant}
\title{Summary of the plant}
\usage{
\method{summary}{plant}(obj)
}
\arguments{
\item{obj}{of class node}
}
\description{
Summary of the plant
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}

