% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{totalLength.root}
\alias{totalLength.root}
\title{Compute the length of the root and its children based on the coordinates of its nodes}
\usage{
\method{totalLength}{root}(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
the total lenght of the root and children
}
\description{
Compute the length of the root and its children based on the coordinates of its nodes
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

