% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xrange.plant}
\alias{xrange.plant}
\title{Compute the xrange of the plant}
\usage{
\method{xrange}{plant}(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
c(x1,x2) where x1 and x2 are the x limits of the plant
}
\description{
Compute the xrange of the plant
}
\keyword{rsml}

