% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xrange.root}
\alias{xrange.root}
\title{Compute the xrange of the root}
\usage{
\method{xrange}{root}(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
c(x1,x2) where x1 and x2 are the x limits of the root
}
\description{
Compute the xrange of the root
}
\keyword{rsml}

