% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xrangePlant}
\alias{xrangePlant}
\title{Compute the xrange of the plant}
\usage{
xrangePlant(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
c(x1,x2) where x1 and x2 are the x limits of the plant
}
\description{
Compute the xrange of the plant
}
\examples{
data(lupin)
xrangePlant(lupin)
}
\keyword{rsml}

