% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{yrange.plant}
\alias{yrange.plant}
\title{Compute the yrange of the plant}
\usage{
\method{yrange}{plant}(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
c(y1,y2) where y1 and y2 are the y limits of the plant
}
\description{
Compute the yrange of the plant
}
\keyword{rsml}

