% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{yrange.root}
\alias{yrange.root}
\title{Compute the xrange of the root}
\usage{
\method{yrange}{root}(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
c(y1,y2) where y1 and y2 are the y limits of the root
}
\description{
Compute the xrange of the root
}
\keyword{rsml}

