% Generated by roxygen2 (4.0.2): do not edit by hand
\name{yrangePlant}
\alias{yrangePlant}
\title{Compute the yrange of the plant}
\usage{
yrangePlant(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
c(y1,y2) where y1 and y2 are the y limits of the plant
}
\description{
Compute the yrange of the plant
}
\examples{
data(lupin)
yrangePlant(lupin)
}
\keyword{rsml}

