% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{zrange.plant}
\alias{zrange.plant}
\title{Compute the zrange of the plant}
\usage{
\method{zrange}{plant}(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
c(z1,z2) where z1 and z2 are the z limits of the plant
}
\description{
Compute the zrange of the plant
}
\keyword{rsml}

