% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zrangePlant}
\alias{zrangePlant}
\title{Compute the zrange of the plant}
\usage{
zrangePlant(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
c(z1,z2) where z1 and z2 are the z limits of the plant
}
\description{
Compute the zrange of the plant
}
\examples{
data(lupin)
zrangePlant(lupin)
}
\keyword{rsml}

