% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{addNodeToRoot}
\alias{addNodeToRoot}
\title{Add a node to an existing root}
\usage{
addNodeToRoot(ro, no)
}
\arguments{
\item{ro}{= the current root}

\item{no}{= the current node}
}
\value{
the root, with the added node
}
\description{
Add a node to an existing root
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

