% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addRootToPlant}
\alias{addRootToPlant}
\title{Add a root to an existing plant. Returns the plant with the added root}
\usage{
addRootToPlant(pl, ro)
}
\arguments{
\item{pl}{the plant to add the root to}

\item{ro}{the root object to add to the plant.}
}
\value{
the new plant, with the added root
}
\description{
Add a root to an existing plant. Returns the plant with the added root
}
\examples{
data(lupin)
r <- root()
lupin <- addRootToPlant(lupin, r)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

