% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{getInsertionPosition}
\alias{getInsertionPosition}
\title{Compute the insertion of the root on its parent}
\usage{
getInsertionPosition(parent, current)
}
\arguments{
\item{parent}{= the parent root}

\item{current}{= the current root}
}
\value{
the insertion position
}
\description{
Compute the insertion of the root on its parent
}

