% Generated by roxygen2 (4.0.2): do not edit by hand
\name{length.plant}
\alias{length.plant}
\title{Compute the length of the root based on the coordinates of its nodes}
\usage{
\method{length}{plant}(x)
}
\arguments{
\item{x}{object of class plant}
}
\value{
the total length of the plant roots
}
\description{
Compute the length of the root based on the coordinates of its nodes
}
\examples{
data(lupin)
length(lupin)
}
\keyword{rsml}

