% Generated by roxygen2 (4.0.2): do not edit by hand
\name{length.root}
\alias{length.root}
\title{Compute the length of the root based on the coordinatres of its nodes}
\usage{
\method{length}{root}(x)
}
\arguments{
\item{x}{object of class root}
}
\value{
the length of the root
}
\description{
Compute the length of the root based on the coordinatres of its nodes
}
\examples{
data(lupin)
r <- lupin$roots[[1]]
length(r)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

