% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_functions.R
\name{meanInsertionAnglePlant}
\alias{meanInsertionAnglePlant}
\title{Compute the mean insertion angle of all the laterals in the plant}
\usage{
meanInsertionAnglePlant(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the mean insertion angle of the root system
}
\description{
Compute the mean insertion angle of all the laterals in the plant
}
\examples{
data(lupin)
meanInsertionAnglePlant(lupin)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}
