% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_functions.R
\name{meanInterbranchPlant}
\alias{meanInterbranchPlant}
\title{Compute the mean interbranch distance of all the primary roots in the image}
\usage{
meanInterbranchPlant(obj, allroot = F)
}
\arguments{
\item{obj}{of class plant}

\item{allroot}{if true, compute the interbanch distance on the whole root}
}
\value{
the mean interbranch distance of the root system
}
\description{
Compute the mean interbranch distance of all the primary roots in the image
}
\examples{
data(lupin)
meanInterbranchPlant(lupin)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}
