% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_functions.R
\name{nLatRoot}
\alias{nLatRoot}
\title{Compute the number of lateral roots in the plant}
\usage{
nLatRoot(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the number of lateral root in the plant
}
\description{
Compute the number of lateral roots in the plant
}
\examples{
data(lupin)
nLatRoot(lupin)
}
\keyword{rsml}
