% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nNode}
\alias{nNode}
\title{Get the number of nodes in a root}
\usage{
nNode(obj)
}
\arguments{
\item{obj}{of class root}
}
\value{
the number of nodes in the root
}
\description{
Get the number of nodes in a root
}
\examples{
data(lupin)
r <- lupin$roots[[1]]
nNode(r)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

