% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_functions.R
\name{nRoot}
\alias{nRoot}
\title{Compute the total number of roots in the plant}
\usage{
nRoot(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the number of root in the plant
}
\description{
Compute the total number of roots in the plant
}
\examples{
data(lupin)
nRoot(lupin)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}
