% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plant}
\alias{plant}
\title{Plant object, containing a root system, composed of roots}
\usage{
plant(roots = NULL)
}
\arguments{
\item{roots}{the root object contained in the plant. Can be null and incremented afterward}
}
\value{
the plant
}
\description{
Plant object, containing a root system, composed of roots
}
\examples{
pl <- plant()
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
\keyword{rsml}

