% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.plant}
\alias{plot.plant}
\title{Plot the root system}
\usage{
\method{plot}{plant}(x, threed = F, ...)
}
\arguments{
\item{x}{object of class plant}

\item{threed}{make a 3D plot for the plant}

\item{...}{plot options}
}
\value{
null
}
\description{
Plot the root system
}
\examples{
# Plot 2D plant
data(lupin)
plot(lupin, threed=FALSE)

# Plot 3D plant
data(anagallis)
plot(anagallis, threed=TRUE)
}
\keyword{rsml}

