% Generated by roxygen2 (4.0.2): do not edit by hand
\name{primLength}
\alias{primLength}
\title{Compute the length of the primary root based on the coordinates of its nodes}
\usage{
primLength(obj)
}
\arguments{
\item{obj}{of class plant}
}
\value{
the total length of the primary roots
}
\description{
Compute the length of the primary root based on the coordinates of its nodes
}
\examples{
data(lupin)
primLength(lupin)
}
\keyword{rsml}

