% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_functions.R
\name{print.plant}
\alias{print.plant}
\title{Print the plant}
\usage{
\method{print}{plant}(x, ...)
}
\arguments{
\item{x}{object of class node}

\item{...}{print options}
}
\description{
Print the plant
}
\examples{
data(lupin)
print(lupin)
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
