% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rsmlToList}
\alias{rsmlToList}
\title{Import a single RSML file into a List. Work only if the roots in the rsml file have properties associated with them.}
\usage{
rsmlToList(rsml.path)
}
\arguments{
\item{rsml.path}{The path to the .rsml file. Mandatory.}
}
\description{
Import a single RSML file into a List. Work only if the roots in the rsml file have properties associated with them.
}
\examples{
path <- "http://rootsystemml.github.io/images/examples/arabidopsis-simple.rsml"
pl.list <- rsmlToList(path) # import the file as a list
write.csv(pl.list$processed, "rsml-table.csv") # save it as a csv table
}
\keyword{rsml}

