% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{rsmlToList}
\alias{rsmlToList}
\title{Import a single RSML file into a List. Work only if the roots in the rsml file have properties associated with them.}
\usage{
rsmlToList(rsml.path)
}
\arguments{
\item{rsml.path}{The path to the .rsml file. Mandatory.}
}
\description{
Import a single RSML file into a List. Work only if the roots in the rsml file have properties associated with them.
}
\keyword{rsml}

