% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.plant}
\alias{summary.plant}
\title{Summary of the plant}
\usage{
\method{summary}{plant}(object, ...)
}
\arguments{
\item{object}{object of class node}

\item{...}{summary options}
}
\description{
Summary of the plant
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}

