% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_functions.R
\name{summary.plant}
\alias{summary.plant}
\title{Summary of the plant}
\usage{
\method{summary}{plant}(object, ...)
}
\arguments{
\item{object}{object of class node}

\item{...}{summary options}
}
\description{
Summary of the plant
}
\examples{
data(lupin)
sum.lup <- summary(lupin)
sum.lup$total.length$value # Get total length
}
\author{
Guillaume Lobet - guillaume.lobet(at)ulg.ac.be
}
